wip(UIKIT-1729,AutoSaveIndicator): Поправлено API и stories
