--- packages/components/src/AutoSave/index.tsx
@@ -1,1 +1,0 @@
-export * from './AutoSave';
